(refactor): scoreboard refactor
